Change Index View of Web Application
https://eaflood.atlassian.net/browse/FPD-503

The heading FFC POC Service should replace Mine Support as the main
title.

this card is to:

update heading to `FFC POC Service`

add text to the index view to explain that this is a POC
service and give context to the user journey.

remove the dead links from the home page.

the index view follows GDS design patterns.